% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intsel-package.R
\docType{package}
\name{intsel-package}
\alias{intsel-package}
\title{intsel: Interaction Selection in Logistic Regression}
\description{
Logistic regression model with data-driven screening for significant two-way interactions.
}
\author{
\strong{Maintainer}: Yi Lian \email{yi.lian@pennmedicine.upenn.edu}

Authors:
\itemize{
  \item Tianze Jiao
  \item Guanbo Wang
  \item Archer Y. Yang
}

Other contributors:
\itemize{
  \item Julien Mairal [contributor]
  \item Yuansi Chen [contributor]
}

}
\keyword{internal}
